# src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: Basic Rest API
  description: Basic Rest API description
  version: 0.1.0
servers:
- url: http://localhost:8080
tags: []
paths:
  /jpa/users:
    get:
      tags:
      - jpa
      summary: 사용자 전체 목록
      description: 사용자 전체 목록 조회하기
      operationId: retrieveAllUsers
      parameters:
      - name: Authorization
        in: header
        description: Basic auth credentials
        required: true
        schema:
          type: string
        example: Basic ZGV2OjEyMzQ=
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jpa-users-1638345299'
              examples:
                retrieveAllUsers:
                  value: "[ {\r\n  \"id\" : 1,\r\n  \"name\" : \"dev1\",\r\n  \"joinDate\"\
                    \ : \"2021-06-23T10:24:10\",\r\n  \"ssn\" : \"12345-56789\",\r\
                    \n  \"posts\" : [ {\r\n    \"id\" : 1,\r\n    \"description\"\
                    \ : \"post 1\"\r\n  }, {\r\n    \"id\" : 3,\r\n    \"description\"\
                    \ : \"post 3\"\r\n  } ]\r\n}, {\r\n  \"id\" : 2,\r\n  \"name\"\
                    \ : \"dev2\",\r\n  \"joinDate\" : \"2021-06-23T10:24:10\",\r\n\
                    \  \"ssn\" : \"12345-56789\",\r\n  \"posts\" : [ {\r\n    \"id\"\
                    \ : 2,\r\n    \"description\" : \"post 2\"\r\n  } ]\r\n}, {\r\n\
                    \  \"id\" : 3,\r\n  \"name\" : \"dev3\",\r\n  \"joinDate\" : \"\
                    2021-06-23T10:24:10\",\r\n  \"ssn\" : \"12345-56789\",\r\n  \"\
                    posts\" : [ ]\r\n} ]"
      security:
      - basic: []
  /jpa/users/{id}:
    get:
      tags:
      - jpa
      summary: 사용자 정보
      description: 사용자 정보 조회하기
      operationId: retrieveUser
      parameters:
      - name: id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: Basic auth credentials
        required: true
        schema:
          type: string
        example: Basic ZGV2OjEyMzQ=
      responses:
        "200":
          description: "200"
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/jpa-users-id-944905545'
              examples:
                retrieveUser:
                  value: "{\r\n  \"name\" : \"dev1\",\r\n  \"joinDate\" : \"2021-06-23T10:24:10\"\
                    ,\r\n  \"ssn\" : \"12345-56789\",\r\n  \"_links\" : {\r\n    \"\
                    all-users\" : {\r\n      \"href\" : \"http://localhost:8080/jpa/users\"\
                    \r\n    }\r\n  }\r\n}"
      security:
      - basic: []
components:
  schemas:
    jpa-users-1638345299:
      type: array
      items:
        type: object
        properties:
          joinDate:
            type: string
            description: 사용자 등록일
          name:
            type: string
            description: 사용자 이름
          id:
            type: number
            description: 아이디
          posts:
            type: array
            description: 사용자 포스트 목록
            items:
              type: object
              properties:
                description:
                  type: string
                  description: 사용자 포스트 내용
                id:
                  type: number
                  description: 사용자 포스트 아이디
          ssn:
            type: string
            description: 사용자 주민번호
    jpa-users-id-944905545:
      type: object
      properties:
        joinDate:
          type: string
          description: 사용자 등록일
        _links:
          type: object
          properties:
            all-users:
              type: object
              properties:
                href:
                  type: string
                  description: 사용자 전체 목록 조회하기
        name:
          type: string
          description: 사용자 이름
        ssn:
          type: string
          description: 사용자 주민번호
  securitySchemes:
    basic:
      type: http
      scheme: basic